requirements.txt
django_heroku
gunicorn